move StaticRequestDispatcher class to internal.io.http package
