+++ components/org.wso2.carbon.uis/src/test/java/org/wso2/carbon/uis/internal/http/RequestDispatcherTest.java
@@ -23,7 +23,7 @@ import org.testng.annotations.Test;
 import org.wso2.carbon.uis.api.exception.UISRuntimeException;
 import org.wso2.carbon.uis.api.http.HttpRequest;
 import org.wso2.carbon.uis.api.http.HttpResponse;
-import org.wso2.carbon.uis.internal.io.StaticRequestDispatcher;
+import org.wso2.carbon.uis.internal.io.http.StaticRequestDispatcher;
 
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.mock;
